add interfaces and not working error handler

+++ ApiRest/src/main/java/com/example/demo/model/external/TaxesServiceResponse.java
@@ -2,6 +2,8 @@ package com.example.demo.model.external;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 
 public class TaxesServiceResponse {
 
@@ -10,23 +12,23 @@ public class TaxesServiceResponse {
     @JsonProperty
     private String name;
     @JsonProperty
-    private Double tax;
+    private BigDecimal tax;
 
 
     public TaxesServiceResponse() {
     }
 
-    public TaxesServiceResponse(String timestamp, String name, Double tax) {
+    public TaxesServiceResponse(String timestamp, String name, BigDecimal tax) {
         this.name = name;
         this.timestamp = timestamp;
         this.tax = tax;
     }
 
-    public Double getTax() {
+    public BigDecimal getTax() {
         return tax;
     }
 
-    public void setTax(Double tax) {
+    public void setTax(BigDecimal tax) {
         this.tax = tax;
     }
