+++ ApiRest/src/main/java/com/example/demo/model/services/calculate/CalculateTaxResponse.java
@@ -3,19 +3,21 @@ package com.example.demo.model.services.calculate;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class CalculateTaxResponse {
 
     @JsonProperty
     private String date;
     @JsonProperty
-    private Double tax;
+    private BigDecimal tax;
     @JsonProperty
-    private Double result;
+    private BigDecimal result;
 
     public CalculateTaxResponse() {
     }
 
-    public CalculateTaxResponse(String date, Double tax, Double result) {
+    public CalculateTaxResponse(String date, BigDecimal tax, BigDecimal result) {
         this.date = date;
         this.tax = tax;
         this.result = result;
@@ -29,19 +31,19 @@ public class CalculateTaxResponse {
         this.date = date;
     }
 
-    public Double getTax() {
+    public BigDecimal getTax() {
         return tax;
     }
 
-    public void setTax(Double tax) {
+    public void setTax(BigDecimal tax) {
         this.tax = tax;
     }
 
-    public Double getResult() {
+    public BigDecimal getResult() {
         return result;
     }
 
-    public void setResult(Double result) {
+    public void setResult(BigDecimal result) {
         this.result = result;
     }
